# BackOffice/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ESTG PAW GRUPO 25 - Food Delivery API",
    "description": "API documentation for the ESTG PAW food delivery platform",
    "version": "1.0.0",
    "contact": {
      "name": "ESTG PAW Team - Grupo 25",
      "email": "8230367@estg.ipp.t"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Order": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Order ID"
          },
          "restaurant": {
            "type": "string",
            "description": "Restaurant ID"
          },
          "customer": {
            "type": "string",
            "description": "Customer ID"
          },
          "order_number": {
            "type": "number",
            "description": "Order number"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "dish": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string" },
                    "description": { "type": "string" },
                    "category": { "type": "string" },
                    "image": { "type": "string" }
                  }
                },
                "dose": {
                  "type": "string",
                  "description": "Portion size"
                },
                "price": {
                  "type": "number",
                  "description": "Item price"
                },
                "quantity": {
                  "type": "number",
                  "description": "Quantity ordered"
                }
              }
            }
          },
          "totalPrice": {
            "type": "number",
            "description": "Total order amount"
          },
          "voucherDiscount": {
            "type": "number",
            "description": "Voucher discount applied"
          },
          "appliedVoucher": {
            "type": "string",
            "description": "Applied voucher ID"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "preparing", "delivered", "finished", "canceled"],
            "description": "Order status"
          },
          "type": {
            "type": "string",
            "enum": ["homeDelivery", "takeAway", "eatIn"],
            "description": "Order type"
          },
          "typeOfcancelation": {
            "type": "string",
            "enum": ["customer", "restaurant", "other"],
            "description": "Type of cancellation"
          },
          "motive": {
            "type": "string",
            "description": "Cancellation motive"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Order creation date"
          },
          "accepted_at": {
            "type": "string",
            "format": "date-time",
            "description": "Order acceptance date"
          },
          "ready_at": {
            "type": "string",
            "format": "date-time",
            "description": "Order ready date"
          },
          "finished_at": {
            "type": "string",
            "format": "date-time",
            "description": "Order finish date"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      }
    }
  },
  "paths": {
...
    "/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get user orders",
        "description": "Get all orders for logged in user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Create order",
        "description": "Create a new order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["restaurant", "items"],
                "properties": {
                  "restaurant": {
                    "type": "string",
                    "description": "Restaurant ID"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "dish": {
                          "type": "string",
                          "description": "Dish ID"
                        },
                        "quantity": {
                          "type": "number",
                          "description": "Quantity"
                        }
                      }
                    }
                  },
                  "deliveryAddress": {
                    "type": "string",
                    "description": "Delivery address"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Invalid order data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}